# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fiteam API Documentation",
    "description": "",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/user/savecard": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "사용자의 성향검사 답변을 결과형태로 저장. 질문의 typeA:I, typeB:E 일 때 응답이 4라면-\u003E I:1, E:4",
        "description": "[{E:4, I:1}, {P:2, D:3}...] 이렇게 사용자 응답결과를 RequestBody로 주세요",
        "operationId": "saveTestResult",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTestAnswerRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/question/unauth/test-result": {
      "post": {
        "tags": [
          "character-question-controller"
        ],
        "summary": "로그인 없이 성향검사 결과 연산",
        "description": "질문 결과를 연산하여 캐릭터카드 ID와 4가지 성향 점수를 반환합니다.",
        "operationId": "saveTestResult_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTestAnswerRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SaveTestResultResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/member/profile/{groupMemberId}": {
      "patch": {
        "tags": [
          "group-member-controller"
        ],
        "summary": "그룹멤버 프로필 수정",
        "description": "초대 수락 후에 그룹 멤버가 자신의 프로필 정보를 수정합니다. (수정 안하는 필드는 null로 입력해주세요",
        "operationId": "updateGroupMemberProfile",
        "parameters": [
          {
            "name": "groupMemberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserGroupProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/noti/{id}": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "알림 상세 조회 및 읽음 처리",
        "description": "알림 ID로 특정 알림을 읽음 처리하고 반환합니다.",
        "operationId": "readNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotifyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notification-controller"
        ],
        "summary": "알림 삭제",
        "description": "알림 ID로 특정 알림을 삭제합니다.",
        "operationId": "deleteNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/member/{userId}/profile": {
      "get": {
        "tags": [
          "group-member-controller"
        ],
        "summary": "다른 멤버 프로필 조회",
        "description": "같은 그룹의 멤버일 경우 해당 사용자의 프로필을 조회합니다.",
        "operationId": "getOtherMemberProfile",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/member/{groupId}/positions": {
      "get": {
        "tags": [
          "group-member-controller"
        ],
        "summary": "유저가 직무 유형 리스트 조회(PM, DS, FE, BE 등등)",
        "description": "teamMakeType 기준으로 TeamType의 configJson을 파싱하여 직무(position) 리스트를 반환합니다.",
        "operationId": "getGroupPositions",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/member/{groupId}/members": {
      "get": {
        "tags": [
          "group-member-controller"
        ],
        "summary": "그룹 전체 멤버 리스트 조회",
        "description": "그룹에 속한 모든 멤버를 조회합니다.",
        "operationId": "getGroupMembers",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/card/{id}": {
      "get": {
        "tags": [
          "character-card-controller"
        ],
        "description": "id(1~16)에 해당하는 캐릭터 Card 가져오기",
        "operationId": "getCharacterCardById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterCard"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SaveTestAnswerRequestDto": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "description": "질문별 답변 리스트\", example = \"[{\\\"E\\\":4,\\\"I\\\":1}, {\\\"P\\\":3,\"D\":2}]",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        }
      },
...
      "SaveTestResultResponseDto": {
        "type": "object",
        "properties": {
          "cardId": {
            "type": "integer",
            "format": "int32",
            "description": "판단된 캐릭터카드 ID",
            "example": 12
          },
          "numEI": {
            "type": "integer",
            "format": "int32",
            "description": "EI 성향 점수",
            "example": 7
          },
          "numPD": {
            "type": "integer",
            "format": "int32",
            "description": "PD 성향 점수",
            "example": 5
          },
          "numVA": {
            "type": "integer",
            "format": "int32",
            "description": "VA 성향 점수",
            "example": 4
          },
          "numCL": {
            "type": "integer",
            "format": "int32",
            "description": "CL 성향 점수",
            "example": 6
          }
        }
      },
...
      "UserGroupProfileDto": {
        "type": "object",
        "properties": {
          "position": {
            "type": "string",
            "description": "포지션",
            "example": "백엔드 개발자"
          },
          "workHistory": {
            "type": "integer",
            "format": "int32",
            "description": "경력(년 수), 1년 미만은 0으로 표시",
            "example": 2
          },
          "projectGoal": {
            "type": "string",
            "description": "프로젝트 목표",
            "example": "포트폴리오 웹사이트 개발"
          },
          "url": {
            "type": "string",
            "description": "개인 URL",
            "example": "https://portfolio.example.com"
          },
          "introduction": {
            "type": "string",
            "description": "자기소개",
            "example": "열정 넘치는 백엔드 개발자입니다."
          }
        }
      },
...
      "UserNotifyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "알림 ID",
            "example": 1
          },
          "senderType": {
            "type": "string",
            "description": "발신자 타입 (user, admin 등)",
            "example": "user"
          },
          "senderId": {
            "type": "integer",
            "format": "int32",
            "description": "발신자 ID (유저/관리자 등)",
            "example": 2
          },
          "type": {
            "type": "string",
            "description": "알림 유형",
            "example": "group_invite"
          },
          "tableId": {
            "type": "integer",
            "format": "int32",
            "description": "관련 엔티티 ID (그룹 ID, 팀 ID 등)",
            "example": 3
          },
          "content": {
            "type": "string",
            "description": "알림 내용",
            "example": "새로운 팀 초대가 도착했습니다."
          },
          "isRead": {
            "type": "boolean",
            "description": "알림 읽음 여부",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "알림 생성 시간",
            "example": "2025-05-05T14:30:00"
          }
        }
      },
...
      "CharacterCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string",
            "example": "EPIC"
          },
          "name": {
            "type": "string",
            "description": "캐릭터 이름",
            "example": "반짝이는 아이디어 조율 요정"
          },
          "keyword": {
            "type": "string",
            "description": "키워드",
            "example": "아이디어, 조율력, 팀 분위기 업"
          },
          "summary": {
            "type": "string",
            "description": "성향 요약",
            "example": "외향적이고 창의적이며 계획을 중시하고 분위기를 잘 맞춤"
          },
          "teamStrength": {
            "type": "string",
            "description": "팀에서의 강점",
            "example": "아이디어 공유, 팀원 간 중재, 분위기메이커"
          },
          "caution": {
            "type": "string",
            "description": "보완해야 할 점",
            "example": "우유부단, 결정 장애, 방향 설정 고민"
          },
          "bestMatchCode1": {
            "type": "string",
            "description": "가장 잘 맞는 캐릭터 코드1",
            "example": "ENFP"
          },
          "bestMatchReason1": {
            "type": "string",
            "description": "베스트 매칭 이유1",
            "example": "감정적인 접근을 통해 INTJ의 균형을 잡아주는 성격"
          },
          "bestMatchCode2": {
            "type": "string",
            "description": "가장 잘 맞는 캐릭터 코드2",
            "example": "INFJ"
          },
          "bestMatchReason2": {
            "type": "string",
            "description": "베스트 매칭 이유2",
            "example": "가치 지향적인 관점을 공유하여 깊은 유대감을 형성"
          },
          "worstMatchCode1": {
            "type": "string",
            "description": "가장 맞지 않는 캐릭터 코드1",
            "example": "ESFP"
          },
          "worstMatchReason1": {
            "type": "string",
            "description": "워스트 매칭 이유1",
            "example": "즉흥적인 행동과 계획적인 성향의 충돌"
          },
          "worstMatchCode2": {
            "type": "string",
            "description": "가장 맞지 않는 캐릭터 코드2",
            "example": "ISFP"
          },
          "worstMatchReason2": {
            "type": "string",
            "description": "워스트 매칭 이유2",
            "example": "갈등 회피 경향으로 인해 소통이 어려울 수 있음"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}